+++ history.md
@@ -7,6 +7,7 @@
 - the warning about the logger when using a string was a bad idea
 - multipart parameters names should not be escaped
 - remove the cookie escaping introduced by migrating to CGI cookie parsing in 1.5.1
+- add a streamed payload type (patch provided by Caleb Land)
 
 # 1.5.1